Factor out order reducer and actions
