+++ app/controllers/action_plans_controller.rb
@@ -1,0 +1,7 @@
+class ActionPlansController < ApplicationController
+  load_and_authorize_resource
+  respond_to :html, :json
+
+  def index
+  end
+end
