+++ app/frontend/order/order.actions.js
@@ -1,0 +1,9 @@
+export const SET_ORDER = 'SET_ORDER';
+
+export function setOrder(order) {
+  return {
+    type: SET_ORDER,
+    order
+  };
+}
+
