+++ app/views/action_plans/index.html.erb
@@ -1,0 +1,3 @@
+<section role="main">
+  <%= react_app("ActionPlans") %>
+</section>
